nbt::tag_list::init Fix
When nbt::tag_list::init was instantiated with the template parameter Arg=std::string with a C++14 conforming standard library (which includes std::make_unique) compilation would fail as overload resolution could not unambiguously select an overload of make_unique: nbt::make_unique was a candidate due to the fact the code-in-question resided in the nbt namespace, however std::make_unique was also a candidate since at least one argument to the function-in-question was from the std namespace (see: Argument dependent lookup).

Fixed this by qualifying the function name at the call site.